Refactor problem service tests and add new validation logic in problemService

+++ Services/JudgeSystem.Services.Data/ProblemService.cs
@@ -35,18 +35,33 @@
 
 		public async Task Delete(Problem problem)
 		{
+            if(!this.Exists(problem.Id))
+            {
+                throw new EntityNotFoundException();
+            }
+
 			problemRepository.Delete(problem);
 			await problemRepository.SaveChangesAsync();
 		}
 
 		public async Task<Problem> GetById(int id)
 		{
-			return await problemRepository.All().FirstOrDefaultAsync(p => p.Id == id);
+            if (!this.Exists(id))
+            {
+                throw new EntityNotFoundException();
+            }
+
+            return await problemRepository.All().FirstOrDefaultAsync(p => p.Id == id);
 		}
 
 		public async Task<Problem> GetByIdWithTests(int id)
 		{
-			return await problemRepository.All()
+            if (!this.Exists(id))
+            {
+                throw new EntityNotFoundException();
+            }
+
+            return await problemRepository.All()
 				.Include(p => p.Tests)
 				.FirstOrDefaultAsync(p => p.Id == id);
 		}
@@ -59,7 +74,12 @@
 
 		public int GetProblemMaxPoints(int id)
 		{
-			var problem = problemRepository.All().FirstOrDefault(p => p.Id == id);
+            if (!this.Exists(id))
+            {
+                throw new EntityNotFoundException();
+            }
+
+            var problem = problemRepository.All().FirstOrDefault(p => p.Id == id);
 			return problem.MaxPoints;
 		}
 
@@ -89,7 +109,12 @@
 
 		public async Task<Problem> Update(ProblemEditInputModel model)
 		{
-			Problem problem = await GetById(model.Id);
+            if (!this.Exists(model.Id))
+            {
+                throw new EntityNotFoundException();
+            }
+
+            Problem problem = await GetById(model.Id);
 			if(problem == null)
 			{
 				throw new EntityNotFoundException(nameof(problem));
@@ -104,5 +129,10 @@
 			await problemRepository.SaveChangesAsync();
 			return problem;
 		}
-	}
+
+        private bool Exists(int id)
+        {
+            return this.problemRepository.All().Any(x => x.Id == id);
+        }
+    }
 }
